Change authorization via social network when turn off session
