+++ modules/user/client-react/components/LoginView.web.jsx
@@ -1,13 +1,31 @@
-import React from 'react';
+import React, { useEffect } from 'react';
 import PropTypes from 'prop-types';
 import Helmet from 'react-helmet';
 
 import { LayoutCenter, PageLayout, Card, CardGroup, CardTitle, CardText, Button } from '@gqlapp/look-client-react';
+import authentication from '@gqlapp/authentication-client-react';
 import settings from '@gqlapp/config';
 
+import saveTokens from '../helpers/saveTokens';
 import LoginForm from './LoginForm';
 
-const LoginView = ({ onSubmit, t, isRegistered, hideModal }) => {
+const LoginView = ({ onSubmit, t, isRegistered, hideModal, history, client }) => {
+  const {
+    location: { search }
+  } = history;
+
+  useEffect(() => {
+    if (search && search.includes('data')) {
+      checkTokensRedirectToProfile();
+    }
+  }, []);
+
+  const checkTokensRedirectToProfile = async () => {
+    await saveTokens(search, client);
+    await authentication.doLogin(client);
+    history.push('profile');
+  };
+
   const renderMetaData = () => (
     <Helmet
       title={`${settings.app.name} - ${t('login.title')}`}
@@ -63,7 +81,9 @@ LoginView.propTypes = {
   onSubmit: PropTypes.func.isRequired,
   t: PropTypes.func,
   isRegistered: PropTypes.bool,
-  hideModal: PropTypes.func
+  hideModal: PropTypes.func,
+  client: PropTypes.object,
+  history: PropTypes.object
 };
 
 export default LoginView;
