+++ modules/user/server-ts/social/shared.js
@@ -9,7 +9,7 @@ export async function onAuthenticationSuccess(req, res) {
   if (redirectUrl) {
     res.redirect(redirectUrl + (tokens ? '?data=' + JSON.stringify({ tokens }) : ''));
   } else {
-    res.redirect('/profile');
+    res.redirect(`/login${tokens ? '?data=' + JSON.stringify({ tokens }) : ''}`);
   }
 }
